Extract AggregateStateId as a separate outer class.
